+++ internal/wrap/public_options.go
@@ -1,0 +1,7 @@
+package wrap
+
+import "github.com/akm/sql-slog/public"
+
+var (
+	NewOptions = public.NewOptions
+)
